+++ src/main/java/com/example/springdemo/repository/CityRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Long>  {
+public interface CityRepository extends JpaRepository<City, Integer>  {
     City findByName(String name);
 }
